use CreateNewRowInData and copy func to dataprocessing.go
